Move savePlayList to useSpotify hook
